Made bike layer work when map is on route mode.
