+++ onebusaway-android/src/amazon/java/org/onebusaway/android/map/googlemapsv2/VehicleOverlay.java
@@ -530,18 +530,21 @@ public class VehicleOverlay implements AmazonMap.OnInfoWindowClickListener {
 
     @Override
     public void onInfoWindowClick(Marker marker) {
-        // Stop any callbacks to refresh the vehicle marker popup balloons
-        mCustomInfoWindowAdapter.cancelUpdates();
+        if (mMarkerData != null) {
+            // Stop any callbacks to refresh the vehicle marker popup balloons
+            mCustomInfoWindowAdapter.cancelUpdates();
 
-        // Show trip details screen for the vehicle associated with this marker
-        ObaTripStatus status = mMarkerData.getStatusFromMarker(marker);
-
-        if (mController != null && mController.getFocusedStopId() != null) {
-            TripDetailsActivity.start(mActivity, status.getActiveTripId(),
-                    mController.getFocusedStopId(), TripDetailsListFragment.SCROLL_MODE_VEHICLE);
-        } else {
-            TripDetailsActivity.start(mActivity, status.getActiveTripId(),
-                    TripDetailsListFragment.SCROLL_MODE_VEHICLE);
+            // Show trip details screen for the vehicle associated with this marker
+            ObaTripStatus status = mMarkerData.getStatusFromMarker(marker);
+            if (status != null) {
+                if (mController != null && mController.getFocusedStopId() != null) {
+                    TripDetailsActivity.start(mActivity, status.getActiveTripId(),
+                            mController.getFocusedStopId(), TripDetailsListFragment.SCROLL_MODE_VEHICLE);
+                } else {
+                    TripDetailsActivity.start(mActivity, status.getActiveTripId(),
+                            TripDetailsListFragment.SCROLL_MODE_VEHICLE);
+                }
+            }
         }
     }
